refactor: renamed parse_input/types => parse_input/raw_input
