+++ crates/committer_cli/src/parse_input.rs
@@ -1,3 +1,3 @@
 pub mod cast;
+pub mod raw_input;
 pub mod read;
-pub mod types;
