+++ crates/committer_cli/src/parse_input/read.rs
@@ -1,6 +1,6 @@
 use committer::{block_committer::input::Input, storage::errors::DeserializationError};
 
-use crate::parse_input::types::RawInput;
+use crate::parse_input::raw_input::RawInput;
 
 #[cfg(test)]
 #[path = "read_test.rs"]
